app/api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=-
#
aiohttp==3.8.6
    # via openai
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
azure-ai-formrecognizer==3.3.1
    # via -r requirements.in
azure-common==1.1.28
    # via
    #   azure-ai-formrecognizer
    #   azure-search-documents
azure-core==1.29.5
    # via
    #   azure-ai-formrecognizer
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
    #   msrest
azure-functions==1.17.0
    # via -r requirements.in
azure-identity==1.15.0
    # via -r requirements.in
azure-search-documents==11.4.0b6
    # via -r requirements.in
azure-storage-blob==12.18.3
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   msrest
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.1
    # via
    #   aiohttp
    #   requests
colorama==0.4.6
    # via tqdm
cryptography==41.0.5
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
et-xmlfile==1.1.0
    # via openpyxl
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
isodate==0.6.1
    # via
    #   azure-search-documents
    #   azure-storage-blob
    #   msrest
msal==1.24.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
msrest==0.7.1
    # via azure-ai-formrecognizer
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.26.1
    # via
    #   openai
    #   pandas
    #   pandas-stubs
oauthlib==3.2.2
    # via requests-oauthlib
openai[datalib]==0.28.1
    # via -r requirements.in
openpyxl==3.1.2
    # via openai
pandas==2.1.2
    # via openai
pandas-stubs==2.1.1.230928
    # via openai
portalocker==2.8.2
    # via msal-extensions
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
pypdf==3.16.4
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
requests==2.31.0
    # via
    #   azure-core
    #   msal
    #   msrest
    #   openai
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via msrest
six==1.16.0
    # via
    #   azure-core
    #   isodate
    #   python-dateutil
tenacity==8.2.3
    # via -r requirements.in
tqdm==4.66.1
    # via openai
types-pytz==2023.3.1.1
    # via pandas-stubs
typing-extensions==4.8.0
    # via
    #   azure-ai-formrecognizer
    #   azure-core
    #   azure-storage-blob
tzdata==2023.3
    # via pandas
urllib3==2.0.7
    # via requests
yarl==1.9.2
    # via aiohttp
